refactor: clean some variable names
